requirements.txt
asyncio
aiofiles
loguru
python-dotenv
retry
openpyxl
urllib3
PyExecJS
requests
argparse
websockets
beautifulsoup4
protobuf
websocket-client
bs4
playwright
qrcode